hash and write file operation

+++ src/cat_file.rs
@@ -41,10 +41,10 @@ where
     reader.read_until(0, &mut buffer)?;
     buffer.pop();
 
-    let size = String::from_utf8(buffer.clone())?.parse::<usize>()?;
+    let size = String::from_utf8(buffer.clone())?.parse::<u64>()?;
 
     let actual_size = std::io::copy(&mut reader, &mut stdout())?;
-    if actual_size != size as u64 {
+    if actual_size != size {
         return Err(anyhow!(
             "Incorrect content length, expected {} but was {}",
             size,
